Moves Stall Fetcher to Services Directory
