+++ src/lib/services/stall_fetcher.ts
